+++ test/OpenTelemetry.Exporter.Geneva.Tests/ConnectionStringBuilderTests.cs
@@ -240,4 +240,19 @@ public class ConnectionStringBuilderTests
         builder = new ConnectionStringBuilder("PrivatePreviewEnableAFDCorrelationIdEnrichment=false");
         Assert.False(builder.PrivatePreviewEnableAFDCorrelationIdEnrichment);
     }
+
+    [Fact]
+    public void ConnectionStringBuilder_PrivatePreviewLogMessagePackStringSizeLimit()
+    {
+        var builder = new ConnectionStringBuilder("key1=value1");
+        Assert.Equal(16383, builder.PrivatePreviewLogMessagePackStringSizeLimit);
+        builder = new ConnectionStringBuilder("PrivatePreviewLogMessagePackStringSizeLimit=1024");
+        Assert.Equal(1024, builder.PrivatePreviewLogMessagePackStringSizeLimit);
+        builder = new ConnectionStringBuilder("PrivatePreviewLogMessagePackStringSizeLimit=65360");
+        Assert.Equal(65360, builder.PrivatePreviewLogMessagePackStringSizeLimit);
+        builder = new ConnectionStringBuilder("PrivatePreviewLogMessagePackStringSizeLimit=-1");
+        Assert.Throws<ArgumentOutOfRangeException>(() => builder.PrivatePreviewLogMessagePackStringSizeLimit);
+        builder = new ConnectionStringBuilder("PrivatePreviewLogMessagePackStringSizeLimit=65361");
+        Assert.Throws<ArgumentOutOfRangeException>(() => builder.PrivatePreviewLogMessagePackStringSizeLimit);
+    }
 }